some modifications to dataset code

+++ src/training_modules/metrics.py
@@ -5,5 +5,5 @@ def get_accuracy(logits: torch.Tensor, labels: torch.Tensor) -> torch.Tensor:
 
 def get_rank(logits: torch.Tensor, labels: torch.Tensor) -> torch.Tensor:
     correct_logits = logits.gather(1, labels.unsqueeze(1))
-    rank = (logits >= correct_logits).sum(dim=1).mean()
+    rank = (logits >= correct_logits).to(torch.float).sum(dim=1).mean()
     return rank
